Initial commit
Contains a C++/CX implementation of a Windows Runtime API which can be
used from a Windows Store application to host the Chakra JavaScript
Runtime Hosting API (JsRT).  The Win32 JsRT API is documented at
https://msdn.microsoft.com/en-us/library/dn249673(v=vs.94).aspx

This project is a higher-level API in general than its Win32 counterpart.
For example, the JavaScriptFunction object has direct access to
Function.bind, call, and apply, which you would have to obtain manually
using the Win32 API.

The initial commit is an early version.  It supports generally a subset of
the functionality accessible to the Internet Explorer 11 version of JsRT.
This project will continue to be updated with additional support.